Large Overhaul, re-introducing steerables with the addition of an interactable class
Camera is centered so the midpoint is 0,0 . Units are being re, introduced, the base ui for uni management is set as a templated and the needed functions are in place, just need the code. Steerable will be made to better apply multiple behaviours in one frame. Formation will be rebuilt and moved from /managers to /game and is no longer classed as a manager class. Objectives will be done from the ground up but its a very small component and will not take much time. Working now with getting formations back into the works and with some sort of advanced placement system and then moving into combat